split module into package/sub-modules
