+++ setup.py
@@ -1,24 +1,22 @@
-import re
+import sys
 import os.path
-from setuptools import setup
+from setuptools import setup, find_packages
 
-MODULE_NAME = 'xnatutils'
+PKG_NAME = 'xnatutils'
 
 # Extract version number from module
-with open(os.path.join(os.path.dirname(__file__),
-                       MODULE_NAME + '.py')) as f:
-    contents = f.read()
-found_versions = re.findall(r'__version__ = (.*)', contents)
-if len(found_versions) != 1:
-    raise Exception("Could not extract version number from module file")
-version = found_versions[0]
+
+
+sys.path.insert(0, os.path.join(os.path.dirname(__file__), PKG_NAME))
+from version_ import __version__  # @IgnorePep8 @UnresolvedImport
+sys.path.pop(0)
 
 setup(
-    name='xnatutils',
-    version=version,
+    name=PKG_NAME,
+    version=__version__,
     author='Tom G. Close',
     author_email='tom.g.close@gmail.com',
-    py_modules=[MODULE_NAME],
+    packages=find_packages(),
     scripts=[os.path.join('scripts', 'xnat-ls'),
              os.path.join('scripts', 'xnat-get'),
              os.path.join('scripts', 'xnat-put'),
@@ -28,7 +26,7 @@ setup(
     license='The MIT License (MIT)',
     description=(
         'A collection of scripts for downloading/uploading and listing '
-        'data from MBI-XNAT'),
+        'data from XNAT repositories.'),
     long_description=open('README.rst').read(),
     install_requires=['xnat>=0.3',
                       'progressbar2>=3.16.0',
