+++ xnatutils/__init__.py
@@ -1,0 +1,6 @@
+from version_ import __version__
+from .base import connect
+from .get import get, varget
+from .put import put, varput
+from .ls import ls
+from .misc import rename
