+++ xnatutils/version_.py
@@ -1,0 +1,2 @@
+
+__version__ = '0.3.1'
